kloud/amazon: add block device support
