+++ go/src/github.com/koding/kloud/provider/amazon/client.go
@@ -7,38 +7,11 @@ import (
 
 	aws "github.com/koding/kloud/api/amazon"
 	"github.com/koding/kloud/machinestate"
-	"github.com/koding/kloud/packer"
 	"github.com/koding/kloud/protocol"
-	"github.com/koding/kloud/utils"
 	"github.com/koding/kloud/waitstate"
 	"github.com/koding/logging"
 	"github.com/mitchellh/goamz/ec2"
 )
-
-// Builder to be used for automatic AMI building with packer
-type ImageBuilder struct {
-	// Credentials
-	AccessKey string `packer:"access_key"`
-	SecretKey string `packer:"secret_key"`
-
-	// Name of AMI to create
-	AmiName string `packer:"ami_name"`
-
-	// Type of instance to create AMI for
-	InstanceType string `packer:"instance_type"`
-
-	// Source AMI to build off
-	SourceAmi string `packer:"source_ami"`
-
-	// AWS region to build AMI in
-	Region string `packer:"region"`
-
-	// SSH login username
-	SshUsername string `packer:"ssh_username"`
-
-	// Type of build ("amazon-ebs")
-	Type string `packer:"type"`
-}
 
 type AmazonClient struct {
 	*aws.Amazon
@@ -134,28 +107,6 @@ func (a *AmazonClient) Build(instanceName string) (*protocol.Artifact, error) {
 		SSHPrivateKey: a.Deploy.PrivateKey,
 		SSHUsername:   "", // deploy with root
 	}, nil
-}
-
-// CreateImage creates an image using Packer. It uses aws.Builder
-// data. It returns the image info.
-func (a *AmazonClient) CreateImage(builder *ImageBuilder, provisioner interface{}) (*ec2.Image, error) {
-	data, err := utils.TemplateData(builder, provisioner)
-	if err != nil {
-		return nil, err
-	}
-
-	provider := &packer.Provider{
-		BuildName: "amazon-ebs",
-		Builder:   data,
-	}
-
-	// this is basically a "packer build template.json"
-	if err := provider.Build(); err != nil {
-		return nil, err
-	}
-
-	// return the image result
-	return a.ImageByName(builder.AmiName)
 }
 
 func (a *AmazonClient) DeployKey() (string, error) {
